# Improve `MultiIf` function performance while type is nullable