--- a/src/Functions/multiIf.cpp
+++ b/src/Functions/multiIf.cpp
@@ -238,7 +238,7 @@ public:
         }
 
         const auto & settings = context->getSettingsRef();
-        const WhichDataType which(result_type);
+        const WhichDataType which(removeNullable(result_type));
         bool execute_multiif_columnar
             = settings.allow_execute_multiif_columnar && !contains_short && (which.isInt() || which.isUInt() || which.isFloat());
 
@@ -254,8 +254,17 @@ public:
     if (which.is##TYPE()) \
     { \
         MutableColumnPtr res = ColumnVector<TYPE>::create(rows); \
-        executeInstructionsColumnar<TYPE, INDEX>(instructions, rows, res); \
-        return std::move(res); \
+        if (!result_type->isNullable()) \
+        { \
+            executeInstructionsColumnar<TYPE, INDEX>(instructions, rows, res); \
+            return std::move(res); \
+        } \
+        else \
+        { \
+            MutableColumnPtr null_map = ColumnUInt8::create(rows); \
+            executeInstructionsColumnarForNullable<TYPE, INDEX>(instructions, rows, res, null_map); \
+            return ColumnNullable::create(std::move(res), std::move(null_map)); \
+        } \
     }
 
 #define ENUMERATE_NUMERIC_TYPES(M, INDEX) \
@@ -370,6 +379,47 @@ private:
                     inserts[row_i] += (~condition_null_map[row_i] & (!!condition_nested_data[row_i])) * (i - inserts[row_i]);
                 }
             }
+        }
+    }
+
+    template<typename T, typename S>
+    static void executeInstructionsColumnarForNullable(std::vector<Instruction> & instructions, size_t rows, const MutableColumnPtr & res, const MutableColumnPtr & null_map)
+    {
+        PaddedPODArray<S> inserts(rows, static_cast<S>(instructions.size()));
+        calculateInserts(instructions, rows, inserts);
+
+        PaddedPODArray<T> & res_data = assert_cast<ColumnVector<T> &>(*res).getData();
+        PaddedPODArray<UInt8> & null_map_data = assert_cast<ColumnUInt8 &>(*null_map).getData();
+        std::vector<const ColumnVector<T> *> data_cols(instructions.size());
+        std::vector<const ColumnUInt8 *> null_map_cols(instructions.size());
+        for (size_t i = 0; i < instructions.size(); ++i)
+        {
+            if (instructions[i].source->isNullable())
+            {
+                const ColumnNullable * nullable_col;
+                if (!instructions[i].source_is_constant)
+                    nullable_col = assert_cast<const ColumnNullable *>(instructions[i].source.get());
+                else
+                {
+                    const ColumnPtr data_column = assert_cast<const ColumnConst *>(instructions[i].source.get())->getDataColumnPtr();
+                    nullable_col = assert_cast<const ColumnNullable *>(data_column.get());
+                }
+                null_map_cols[i] = assert_cast<const ColumnUInt8 *>(nullable_col->getNullMapColumnPtr().get());
+                data_cols[i] = assert_cast<const ColumnVector<T> *>(nullable_col->getNestedColumnPtr().get());
+            }
+            else
+            {
+                null_map_cols[i] = nullptr;
+                data_cols[i] = assert_cast<const ColumnVector<T> *>(instructions[i].source.get());
+            }
+        }
+
+        for (size_t row_i = 0; row_i < rows; ++row_i)
+        {
+            auto & instruction = instructions[inserts[row_i]];
+            size_t index = instruction.source_is_constant ? 0 : row_i;
+            res_data[row_i] = data_cols[inserts[row_i]]->getData()[index];
+            null_map_data[row_i] = null_map_cols[inserts[row_i]]->getData()[index];
         }
     }