… tests/performance/multiif.xml
… tests/queries/0_stateless/02918_multif_for_nullable.reference
… tests/queries/0_stateless/02918_multif_for_nullable.sh