<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):
- Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
#### Improve on nullable type
We find in `multiIf` function, when the return column type is `nullable`, then it would not use `executeInstructionsColumnar` method, which may decrease its performance. So we introduce a similar method `executeInstructionsColumnarForNullable` for nullable data type, to improve the function's performance in nullable case

Here, we test its performance below, the test table `test_tbl(d Nullable(Int64))` has about 300000000 rows data range from 1~300000000, and we do the query ` select count(1) from test_tbl where multiIf(d > 2, d-2, d > 1, d-1, d >0, d, 0) > 1 SETTINGS max_threads=1`, test for 3 times

before this pr's modification, test as below
```
1 row in set. Elapsed: 8.121 sec. Processed 300.00 million rows, 2.70 GB (36.94 million rows/s., 332.47 MB/s.)
1 row in set. Elapsed: 7.169 sec. Processed 300.00 million rows, 2.70 GB (41.84 million rows/s., 376.60 MB/s.)
1 row in set. Elapsed: 7.674 sec. Processed 300.00 million rows, 2.70 GB (39.09 million rows/s., 351.83 MB/s.)
```

after this pr's modification, test as below
```
1 row in set. Elapsed: 3.506 sec. Processed 300.00 million rows, 2.70 GB (85.56 million rows/s., 770.08 MB/s.)
1 row in set. Elapsed: 3.664 sec. Processed 300.00 million rows, 2.70 GB (81.88 million rows/s., 736.96 MB/s.)
1 row in set. Elapsed: 3.511 sec. Processed 300.00 million rows, 2.70 GB (85.46 million rows/s., 769.12 MB/s.)
```
### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/