@KevinyhZou 2023-12-18
It seems to use memcpy when the condition skew not improve the performance, so I have remove the memcpy copy logical, and I have reuse `executeInstructionsColumnar` to handle nullable case. @taiyang-li